Add new function(see description)
Added the ability to create devices on command from the user mod. Fixed a bug with access to a device that was not created in DriverEntry. Added the ability to mount a disk, create a file. Simple added
encryption.
Also, the driver was tested for the first time on Windows 10. More checks are required.